--- tests/MeshSendRecvTest.cpp
@@ -1,53 +1,0 @@
-#include "l3ster/comm/DistributeMesh.hpp"
-#include "l3ster/util/ScopeGuards.hpp"
-
-#include "Common.hpp"
-#include "TestDataPath.h"
-
-#include <iostream>
-
-using namespace lstr;
-using namespace lstr::comm;
-using namespace lstr::mesh;
-
-using el_ptr_var = Domain< 1 >::el_univec_t::const_ptr_variant_t;
-bool compareElemsAt(el_ptr_var v1, el_ptr_var v2)
-{
-    constexpr auto compare_contents = [](auto ptr1, auto ptr2) {
-        if constexpr (std::same_as< decltype(ptr1), decltype(ptr2) >)
-            return *ptr1 == *ptr2;
-        return false;
-    };
-    return std::visit(compare_contents, v1, v2);
-}
-
-int main(int argc, char* argv[])
-{
-    L3sterScopeGuard scope_guard{argc, argv};
-    MpiComm          comm{MPI_COMM_WORLD};
-
-    const auto read_mesh = readMesh(L3STER_TESTDATA_ABSPATH(gmsh_ascii4_square.msh), {}, lstr::mesh::gmsh_tag);
-
-    if (comm.getRank() == 0)
-    {
-        auto reqs = std::vector< MpiComm::Request >{};
-        for (int i = 1; i < comm.getSize(); ++i)
-            sendMesh(comm, read_mesh, i, std::back_inserter(reqs));
-        MpiComm::Request::waitAll(reqs);
-    }
-    else
-    {
-        const auto recv_mesh = receiveMesh< 1 >(comm, 0);
-        REQUIRE(std::ranges::equal(read_mesh.getAllNodes(), recv_mesh.getAllNodes()));
-        REQUIRE(std::ranges::equal(read_mesh.getDomainIds(), recv_mesh.getDomainIds()));
-        for (d_id_t domain_id : read_mesh.getDomainIds())
-        {
-            const auto& read_domain = read_mesh.getDomain(domain_id);
-            const auto& recv_domain = recv_mesh.getDomain(domain_id);
-            REQUIRE(read_domain.dim == recv_domain.dim);
-            REQUIRE(read_domain.elements.size() == recv_domain.elements.size());
-            for (size_t i = 0; i != read_domain.elements.size(); ++i)
-                REQUIRE(compareElemsAt(read_domain.elements.at(i), recv_domain.elements.at(i)));
-        }
-    }
-}
